requirements.txt
contourpy==1.3.0
cycler==0.12.1
ecos==2.0.14
fastplot==1.4.0
fonttools==4.54.1
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.9.2
mpmath==1.3.0
numexpr==2.10.1
numpy==2.1.3
osqp==0.6.7.post3
packaging==24.1
pandas==2.2.3
patsy==0.5.6
pillow==11.0.0
prettytable==3.12.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
qdldl==0.1.7.post4
scikit-learn==1.5.2
scikit-survival==0.23.1
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
statsmodels==0.14.4
sympy==1.13.3
threadpoolctl==3.5.0
tqdm==4.66.6
tzdata==2024.2
wcwidth==0.2.13
